refactor(main): DiffAppState tests を state module に移動
state/flow tests を DiffAppState 実装近くへ移し、main.rs から test-only module を削除する。

Refs: #224
